Add ticks to saved state. (#8)
* Add option to add ticks to saved state.

Do some cleanup too.

* Prevent writing ticks when its 0.

This could be the case if the xml doesn't have the ticks field yet, so it would be the default 0 if the field was missing.

* Change to changing snapshot stats instead

Inspired by JeFis question about how its implemented. Instead of chaging all time stats we change snapshot stats, as it leaves all time stats intact, but also makes for a more solid approach as times could be negative if an old savestate was loaded on a new map. Also all the other changes I talked about to make the diff extra ugly.